Restructure CLI commands for simpler organization
Move all subcommands from cmd/binst/ to cmd/ directory and extract rootCmd
to cmd/root.go, making main.go minimal. This creates a cleaner structure
where main.go only handles build variables and calls cmd.Execute().

Changes:
- Move cmd/binst/{init,gen,embed_checksums}.go to cmd/
- Create cmd/root.go with rootCmd and all setup logic
- Simplify main.go to just 19 lines
- All subcommands now in single cmd package for easier sharing

Benefits:
- Simpler directory structure without nested packages
- main.go is now extremely minimal and focused
- Easier to run with go run ./cmd/binst
- Shared utilities like resolveConfigFile accessible to all commands
